Refactor route stop markers and enable automativ popup open.
