+++ src/components/map/CompoundStopMarker.js
@@ -82,7 +82,7 @@ const CompoundStopMarker = decorate(
       if (popupOpen && markerRef.current) {
         markerRef.current.leafletElement.openPopup();
         didAutoOpen.current = true;
-      } else if (didAutoOpen.current) {
+      } else if (didAutoOpen.current && markerRef.current) {
         markerRef.current.leafletElement.closePopup();
       }
     }, [popupOpen]);
