+++ src/components/map/RouteStopsLayer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import flow from "lodash/flow";
-import RouteStopMarker from "./RouteStopMarker";
+import RouteStop from "./RouteStop";
 import {observer} from "mobx-react-lite";
 import StopsByRouteQuery from "../../queries/StopsByRouteQuery";
 import {inject} from "../../helpers/inject";
@@ -36,7 +36,7 @@ const RouteStopsLayer = decorate(
             const isSelected = stop.stopId === selectedStop;
 
             return (
-              <RouteStopMarker
+              <RouteStop
                 key={`stop_marker_${stop.stopId}`}
                 selected={isSelected}
                 firstTerminal={isFirst}
